fix invalid cast error

+++ ferrischat_webserver/src/users/get_user.rs
@@ -64,7 +64,7 @@ pub async fn get_user(req: HttpRequest, auth: crate::Authorization) -> impl Resp
                 } else {
                     None
                 },
-                flags: user.flags as u64,
+                flags: user.flags,
             }),
             None => HttpResponse::NotFound().json(NotFoundJson {
                 message: "User Not Found".to_string(),
